[Refactor]: 커뮤니티 게시글 생성 API 리팩토링
PostSaveRequest로 받는 데이터 validation 추가

Related to: #559
